refactor(crawler): split agent.rs into agent/ module directory
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
